Add start, duplicate and snooze in overview
