+++ app/src/main/java/io/ipoli/android/quest/events/DuplicateQuestRequestEvent.java
@@ -2,6 +2,7 @@ package io.ipoli.android.quest.events;
 
 import java.util.Date;
 
+import io.ipoli.android.app.events.EventSource;
 import io.ipoli.android.quest.data.Quest;
 
 /**
@@ -11,14 +12,17 @@ import io.ipoli.android.quest.data.Quest;
 public class DuplicateQuestRequestEvent {
     public final Quest quest;
     public final Date date;
+    public final EventSource source;
 
-    public DuplicateQuestRequestEvent(Quest quest) {
+    public DuplicateQuestRequestEvent(Quest quest, EventSource source) {
         this.quest = quest;
         this.date = null;
+        this.source = source;
     }
 
-    public DuplicateQuestRequestEvent(Quest quest, Date date) {
+    public DuplicateQuestRequestEvent(Quest quest, Date date, EventSource source) {
         this.quest = quest;
         this.date = date;
+        this.source = source;
     }
 }
