+++ app/src/main/java/io/ipoli/android/quest/ui/menus/DuplicateDateItem.java
@@ -1,0 +1,17 @@
+package io.ipoli.android.quest.ui.menus;
+
+import java.util.Date;
+
+/**
+ * Created by Polina Zhelyazkova <polina@ipoli.io>
+ * on 8/14/16.
+ */
+public class DuplicateDateItem {
+    public String title;
+    public Date date;
+
+    public DuplicateDateItem(String title, Date date) {
+        this.title = title;
+        this.date = date;
+    }
+}
